chore: rename core to pkg
