+++ packages/pkg/index.html
@@ -9,7 +9,7 @@
 
 <body>
   <div id="root"></div>
-  <script type="module" src="./playground/main.tsx"></script>
+  <script type="module" src="./ui/main.tsx"></script>
 </body>
 
 </html>
