+++ packages/pkg/src/index.ts
@@ -1,9 +1,9 @@
-import { LoudnessProcessor } from "#processors/loudness-processor";
 import type {
   LoudnessMeasurements,
   LoudnessProcessorOptions,
   LoudnessSnapshot,
-} from "#types";
+} from "#common/types";
+import { LoudnessProcessor } from "#modules/loudness-processor";
 
 registerProcessor("loudness-processor", LoudnessProcessor);
