+++ packages/pkg/src/modules/finite-impulse-response-filter.ts
