+++ packages/pkg/vite.config.ts
